requirements.txt
python-telegram-bot==20.7
paramiko==3.4.0
celery==5.3.6
redis==5.0.1
SQLAlchemy==2.0.25
cryptography==41.0.8
python-dotenv==1.0.0
asyncssh==2.14.2
aiofiles==23.2.1
flower==2.0.1
kombu==5.3.4
